requirements.txt
codebleu
dashscope
numpy
qianfan
Requests
volcengine
websocket_client
zhipuai
